define employee reducer, refactor auth reducer into its own file
