+++ src/actions/employeeActions.js
@@ -1,0 +1,6 @@
+import EMPLOYEE_UPDATE from './types';
+
+export const employeeUpdate = ({ prop, value }) => ({
+  type: EMPLOYEE_UPDATE,
+  payload: { prop, value }
+});
