+++ src/actions/types.js
@@ -3,3 +3,5 @@ export const PASSWORD_CHANGED = 'PASSWORD_CHANGED';
 export const LOGIN_USER_SUCCESS = 'LOGIN_USER_SUCCESS';
 export const LOGIN_USER_FAIL = 'LOGIN_USER_FAIL';
 export const LOGIN_USER = 'LOGIN_USER';
+
+export const EMPLOYEE_UPDATE = 'EMPLOYEE_UPDATE';
