+++ src/reducers/index.js
@@ -1,6 +1,8 @@
 import { combineReducers } from 'redux';
 import login from './login';
+import employeeForm from './employeeForm';
 
 export default combineReducers({
-  login
+  login,
+  employeeForm
 });
